Move `DatagramDispatcher` to dedicated module
